Move RequestExtension from js_generic to public includes
