+++ include/ccf/js/named_auth_policies.h
